+++ src/main/java/fr/gamecreep/fundraiserfusion/donations/DonationCache.java
@@ -1,17 +1,17 @@
-package fr.gamecreep.streamlabsdonations.donations;
+package fr.gamecreep.fundraiserfusion.donations;
 
-import fr.gamecreep.streamlabsdonations.StreamLabsDonations;
-import fr.gamecreep.streamlabsdonations.donations.entities.Donation;
+import fr.gamecreep.fundraiserfusion.FundraiserFusion;
+import fr.gamecreep.fundraiserfusion.donations.entities.Donation;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
 public class DonationCache {
-    private final StreamLabsDonations plugin;
+    private final FundraiserFusion plugin;
     private final List<Donation> totalDonations;
 
-    public DonationCache(StreamLabsDonations plugin) {
+    public DonationCache(FundraiserFusion plugin) {
         this.plugin = plugin;
         totalDonations = new ArrayList<>();
     }
