+++ src/main/java/fr/gamecreep/fundraiserfusion/donations/events/Donation50.java
@@ -1,4 +1,4 @@
-package fr.gamecreep.streamlabsdonations.donations.events;
+package fr.gamecreep.fundraiserfusion.donations.events;
 
 import org.bukkit.entity.Player;
