--- src/main/resources/config.yml
@@ -1,2 +1,0 @@
-streamlabs:
-  accessToken: ""
